requirements.txt
notion-client==2.2.1
anthropic==0.40.0
groq==0.12.0
httpx==0.27.2
python-dotenv==1.0.1
schedule==1.2.2
psutil==6.1.0
requests==2.32.3
paramiko==3.5.0
pydantic==2.9.2
rich==13.9.4
tenacity==9.0.0
